Regenerated images/comedi_logo.png from images/comedi_logo.svg
Used the following commands to regenerate comedi_logo.png so that it
matches the new comedi_logo.svg file at width 417:

   inkscape -z -e comedi_logo.png -w 417 comedi_logo.svg
   optipng comedi_logo.png

Optipng reduces the size of the file.  In this case, it converted it to
an 8-bit palleted colour image with alpha.

Also, remove the GIMP format images/comedi_logo.xcf as it is no longer
relevant.